feat: osmosis graphql
