+++ src/lib/app-provider/query-client/index.ts
@@ -4,7 +4,7 @@ export const GRAPH_URL: Record<string, string> = {
   /**
    * Revisit graphql for terra2 mainnet and osmosis mainnet
    */
-  osmosis: "https://osmosis-mainnet-graphql.alleslabs.dev/v1/graphql",
+  osmosis: "https://osmosis-mainnet-graphql2.alleslabs.dev/v1/graphql",
   osmosistestnet: "https://osmosis-testnet-graphql2.alleslabs.dev/v1/graphql",
   terra2testnet: "https://terra-testnet-graphql.alleslabs.dev/v1/graphql",
 };
