+++ src/lib/types/asset.ts
@@ -1,0 +1,12 @@
+import type Big from "big.js";
+
+import type { Option, Token, U, USD } from "lib/types";
+
+export interface TokenWithValue {
+  denom: string;
+  amount: U<Token<Big>>;
+  symbol: Option<string>;
+  logo: Option<string>;
+  precision: Option<number>;
+  value: Option<USD<Big>>;
+}
