+++ src/lib/types/index.ts
@@ -10,3 +10,4 @@ export * from "./projects";
 export * from "./addrs";
 export * from "./proposal";
 export * from "./validator";
+export * from "./asset";
