+++ src/lib/utils/assetValue.ts
@@ -1,7 +1,16 @@
 import type { BigSource, Big } from "big.js";
 import big from "big.js";
 
-import type { Balance, BalanceWithAssetInfo, Token, U, USD } from "lib/types";
+import type {
+  AssetInfo,
+  Balance,
+  BalanceWithAssetInfo,
+  Option,
+  Token,
+  TokenWithValue,
+  U,
+  USD,
+} from "lib/types";
 
 import { toToken } from "./formatter";
 
@@ -26,3 +35,36 @@ export const calTotalValue = (assets: BalanceWithAssetInfo[]): USD<Big> =>
       acc.add(calAssetValueWithPrecision(curr.balance)) as USD<Big>,
     big(0) as USD<Big>
   );
+
+export const coinToTokenWithValue = (
+  denom: string,
+  amount: string,
+  assetInfo: Option<AssetInfo>
+): TokenWithValue => {
+  const tokenAmount = big(amount) as U<Token<Big>>;
+  return {
+    denom,
+    amount: tokenAmount,
+    symbol: assetInfo?.symbol,
+    logo: assetInfo?.logo,
+    precision: assetInfo?.precision,
+    value: assetInfo
+      ? calculateAssetValue(
+          toToken(tokenAmount, assetInfo.precision),
+          assetInfo.price as USD<number>
+        )
+      : undefined,
+  };
+};
+
+export const addTokenWithValue = (
+  oldTotal: Option<TokenWithValue>,
+  token: TokenWithValue
+): TokenWithValue =>
+  !oldTotal
+    ? token
+    : {
+        ...oldTotal,
+        amount: oldTotal.amount.add(token.amount) as U<Token<Big>>,
+        value: oldTotal.value?.add(token.value ?? 0) as USD<Big>,
+      };
